chore: move user, raffle, participant and formQuestion interfaces to interfaces/ directory
