+++ src/auth/auth.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { User, UserService } from 'src/user/user.service';
+import { UserService } from 'src/user/user.service';
+import { User } from 'src/interfaces/user';
 import { jwtConstants } from './constants';
 
 @Injectable()
