+++ src/interfaces/formQuestion.ts
@@ -1,0 +1,6 @@
+export interface FormQuestion {
+  type: string,
+  title: string,
+  description: string,
+  very_important_field: object,
+}
