+++ src/interfaces/participant.ts
@@ -1,0 +1,5 @@
+export interface Participant {
+    participantid: string;
+    info: object;
+    raffleid: string;
+}
