+++ src/interfaces/raffle.ts
@@ -1,0 +1,8 @@
+import { FormQuestion } from "src/interfaces/formQuestion";
+
+export interface Raffle {
+  raffleid: string,
+  rafflename: string,
+  form: FormQuestion[],
+  authorid: string,
+}
