+++ src/interfaces/user.ts
@@ -1,0 +1,7 @@
+export interface User {
+  userid: string;
+  username: string;
+  email: string;
+  password: string;
+  created_at: string;
+};
